+++ src/Bridge/Symfony/Blacklister.php
@@ -1,0 +1,23 @@
+<?php
+
+/*
+ * This file is part of the SeoOverride project.
+ *
+ * (c) JoliCode <coucou@jolicode.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Joli\SeoOverride\Bridge\Symfony;
+
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+
+/**
+ * A blacklister prevents SeoOverride to run matchers on some requests.
+ */
+interface Blacklister
+{
+    public function isBlacklisted(Request $request, Response $response): bool;
+}